Move hand pose code into HandPose file
