+++ HandPose.h
@@ -11,4 +11,12 @@ namespace F4VRBody {
 	extern std::map<std::string, bool> handPapyrusHasControl;
 
 	void initHandPoses(bool inPowerArmor);
+	
+	void setFingerPositionScalar(bool isLeft, float thumb, float index, float middle, float ring, float pinky);
+	void restoreFingerPoseControl(bool isLeft);
+	
+	void setPipboyHandPose();
+	void disablePipboyHandPose();
+	void setConfigModeHandPose();
+	void disableConfigModePose();
 }
