+++ Pipboy.cpp
@@ -3,6 +3,7 @@
 #include "Config.h"
 #include "utils.h"
 #include "VR.h"
+#include "HandPose.h"
 #include "F4VRBody.h"
 
 #include <chrono>
@@ -176,7 +177,7 @@ namespace F4VRBody {
 					g_boneSpheres->drawWeapon(); // draw weapon as we no longer need primary trigger as an input.
 					_weaponStateDetected = false;
 				}
-				_skelly->disablePipboyHandPose();
+				disablePipboyHandPose();
 				_skelly->getPlayerNodes()->PipboyRoot_nif_only_node->m_localTransform.scale = 0.0;
 				_MESSAGE("Disabling Pipboy with button");
 				_stickyoffpip = true;
@@ -209,7 +210,7 @@ namespace F4VRBody {
 					}
 				}
 				turnPipBoyOn();
-				_skelly->setPipboyHandPose();
+				setPipboyHandPose();
 				_isOperatingPipboy = true;
 				_MESSAGE("Enabling Pipboy with button");
 				_stickybpip = false;
@@ -232,7 +233,7 @@ namespace F4VRBody {
 					g_boneSpheres->drawWeapon(); // draw weapon as we no longer need primary trigger as an input.
 					_weaponStateDetected = false;
 				}
-				_skelly->disablePipboyHandPose();
+				disablePipboyHandPose();
 				_isOperatingPipboy = false;
 				//		_MESSAGE("Disabling PipBoy due to inactivity for %d more than %d ms", timeElapsed, g_config->pipBoyOffDelay);
 			}
@@ -244,7 +245,7 @@ namespace F4VRBody {
 					g_boneSpheres->drawWeapon(); // draw weapon as we no longer need primary trigger as an input.
 					_weaponStateDetected = false;
 				}
-				_skelly->disablePipboyHandPose();
+				disablePipboyHandPose();
 				_isOperatingPipboy = false;
 				//		_MESSAGE("Disabling PipBoy due to movement when not looking at pipboy. input: (%f, %f)", axis_state.x, axis_state.y);
 			}
@@ -270,7 +271,7 @@ namespace F4VRBody {
 						}
 					}
 					turnPipBoyOn();
-					_skelly->setPipboyHandPose();
+					setPipboyHandPose();
 					_isOperatingPipboy = true;
 					_startedLookingAtPip = 0;
 				}
@@ -480,11 +481,11 @@ namespace F4VRBody {
 						_weaponStateDetected = true;
 						g_boneSpheres->holsterWeapon();
 					}
-					_skelly->setPipboyHandPose();
+					setPipboyHandPose();
 				}
 				if ((distance > g_config->pipboyDetectionRange) && _isOperatingPipboy && !_pipboyStatus) { // Restores Weapon and releases hand pose
 					_isOperatingPipboy = false;
-					_skelly->disablePipboyHandPose();
+					disablePipboyHandPose();
 					if (_isWeaponinHand) {
 						_weaponStateDetected = false;
 						g_boneSpheres->drawWeapon();
@@ -492,7 +493,7 @@ namespace F4VRBody {
 				}
 			}
 			else if (!isLookingAtPipBoy() && _isOperatingPipboy && !_pipboyStatus) { // Catches if you're not looking at the pipboy when your hand moves outside of the control area and restores weapon / releases hand pose							
-				_skelly->disablePipboyHandPose();
+				disablePipboyHandPose();
 				for (int i = 0; i < 7; i++) {  // Remove any stuck helper orbs if Pipboy times out for any reason.
 					NiAVObject* orb = g_config->leftHandedPipBoy ? _skelly->getRightArm().forearm3->GetObjectByName(&orbNames[i]) : _skelly->getLeftArm().forearm3->GetObjectByName(&orbNames[i]);
 					if (orb != nullptr) {
